Refactor DICOM viewer state management; improve folder and file selection handling and enhance user notifications
